requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.9.0
    # via flask
click==8.2.1
    # via flask
colorama==0.4.6
    # via click
flask==3.0.0
    # via
    #   -r requirements.in
    #   flask-login
    #   flask-sqlalchemy
    #   flask-wtf
flask-login==0.6.3
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via -r requirements.in
flask-wtf==1.2.1
    # via -r requirements.in
itsdangerous==2.2.0
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
    #   wtforms
python-dateutil==2.8.2
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.36
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
typing-extensions==4.15.0
    # via sqlalchemy
werkzeug==3.0.1
    # via
    #   -r requirements.in
    #   flask
    #   flask-login
wtforms==3.1.1
    # via
    #   -r requirements.in
    #   flask-wtf
